Package extension from xtask Not tested on Windows
