+++ xtask/src/main.rs
@@ -6,6 +6,7 @@ fn main() {
 
     let res = match task.as_deref() {
         Some("codegen") => xtask::do_codegen(),
+        Some("pack-ext") => xtask::do_package(),
         _ => {
             show_help();
             Ok(())
@@ -22,6 +23,7 @@ fn show_help() {
     println!(
         r#"Available tasks:
 
-codegen         Performs necessary code generation"#
+codegen         Performs necessary code generation
+pack-ext        Packages the extension"#
     );
 }
